move random_identifier() to mrjob.util and test it
