+++ CHANGES.txt
@@ -29,6 +29,7 @@ v0.5.0, 2015-??-?? -- ???
      * hdfs_scratch_dir option is now hadoop_tmp_dir (#318)
  * removed mrjob.compat.get_jobconf_value() (use jobconf_from_env())
  * mrjob.util:
+   * random_identifier() moved here from mrjob.aws
    * renamed buffer_iterator_to_line_iterator() to to_lines()
      * to_lines() no longer appends a newline to data (#819)
    * removed extract_dir_for_tar()
